fix(file-transfer): align with OpenSpec spec and fix test/cross-build issues
- proto: change size_bytes int64 → uint64; server/web.rs: remove as i64 cast
- client/ui: always start file worker; resolve %USERPROFILE%\Downloads as
  default save dir when file_save_dir is not configured
- client/file_handler: fix CF_DIBV5 bV5Height to positive (bottom-up) and
  add row-flip + RGBA→BGRA per-row to match Windows DIB orientation
- client/file_handler: sanitize trailing whitespace/dots only (keep leading dots)
- client/file_handler: conflict numbering format photo(1).jpg (remove space)
- client/file_handler: CF_DIBV5.into() for SetClipboardData u16→u32 (windows-sys 0.61)
- client.example.toml: rename key to file_saved_path, add alias in config.rs,
  add file_download_max_retries and image_clipboard_max_decoded_bytes examples
- server/config.rs: inject upload_dir in test parse helper to fix 15 pre-existing
  test failures

+++ server/src/web.rs
@@ -863,7 +863,7 @@ async fn handle_upload(
                 file_id: file_id.to_string(),
                 file_name: meta.file_name.clone(),
                 mime_type: meta.mime_type.clone(),
-                size_bytes: meta.size_bytes as i64,
+                size_bytes: meta.size_bytes,
                 download_url,
                 sha256: meta.sha256.clone(),
                 transfer_mode: TransferMode::Relay as i32,
